+++ Source/Localization/Text.swift
@@ -1,4 +1,4 @@
-// swiftlint:disable line_length identifier_name nesting
+// swiftlint:disable line_length identifier_name nesting file_length
 
 // provide the types of any new Localizable enums
 // in order to automatically export their strings to the localization files
@@ -23,6 +23,13 @@ extension Text {
             Text.Channel.self,
             Text.Post.self,
             Text.Report.self,
+            Text.Notifications.self,
+            Text.Help.self,
+            Text.Help.Home.self,
+            Text.Help.Discover.self,
+            Text.Help.Notifications.self,
+            Text.Help.Hashtags.self,
+            Text.Help.YourNetwork.self,
         ]
     }
 }
@@ -65,6 +72,7 @@ enum Text: String, Localizable, CaseIterable {
     case startedFollowing = "{{somebody}} started following"
     case stoppedFollowing = "{{somebody}} stopped following"
     case followStats = "Following {{numberOfFollows}} • Followed by {{numberOfFollowers}}"
+    case userOusideNetwork = "This user is outside your network"
     case showMeInDirectory = "Show me in the directory"
     case showMeInUserDirectory = "Show me in the user directory"
     case hideMeFromUserDirectory = "Hide me from the user directory"
@@ -96,8 +104,8 @@ enum Text: String, Localizable, CaseIterable {
     case likesThis = "likes this"
     case dislikesThis = "dislikes this"
 
-    case block = "Block"
-    case blocked = "Blocked"
+    case block = "Ignore"
+    case blocked = "Ignored"
 
     case deleteSecretAndIdentity = "Delete this secret and identity"
 
@@ -129,8 +137,8 @@ enum Text: String, Localizable, CaseIterable {
     case addFriend = "Add friend"
     case removeFriend = "Remove from friends"
 
-    case blockUser = "Block this user"
-    case unblockUser = "Unblock this user"
+    case blockUser = "Ignore this user"
+    case unblockUser = "Unignore this user"
 
     case reportPost = "Report this post"
     case reportUser = "Report this user"
@@ -200,6 +208,7 @@ start using Planetary right away, but:
 """
     case startUsingPlanetary = "start using Planetary"
     case startUsingPlanetaryTitle = "Start Using Planetary"
+    case dismissAndStartUsingPlanetary = "Dismiss and start using Planetary"
     case percentComplete = "complete"
     case beta1Disclaimers =
 """
@@ -410,6 +419,15 @@ extension Text {
     }
 }
 
+// MARK: - Notifications
+
+extension Text {
+
+    enum Notifications: String, Localizable, CaseIterable {
+        case markAllAsRead = "Mark all as read"
+    }
+}
+
 // MARK: - Report
 
 extension Text {
@@ -428,12 +446,12 @@ extension Text {
 extension Text {
 
     enum Blocking: String, Localizable, CaseIterable {
-        case alertTitle = "Are you sure you want to block {{ name }}? You will no longer see each other's content or be able to contact each other."
-        case buttonTitle = "Yes, block {{ name }}"
-        case blockedUsers = "Blocked Users"
-        case footer = "Blocked users cannot see your posts or contact you, and you will need to unblock them before you can see their posts or contact them. It may take some time to see users and content once they have been unblocked."
+        case alertTitle = "Are you sure you want to ignore {{ name }}? You will no longer see each other's content or be able to contact each other. This will be publicly visible."
+        case buttonTitle = "Yes, ignore {{ name }}"
+        case blockedUsers = "Ignored Users"
+        case footer = "Ignored users cannot see your posts or contact you, and you will need to unignore them before you can see their posts or contact them. It may take some time to see users and content once they have been unignored."
         case thisUser = "this user"
-        case usersYouHaveBlocked = "Users that you have blocked"
+        case usersYouHaveBlocked = "Users that you have ignored"
     }
 }
 
@@ -485,5 +503,43 @@ extension Text {
         case postReplied = "%@ replied to your post"
         case feedMentioned = "%@ mentioned you in a post"
         case messageLiked = "%@ liked your post"
+    }
+}
+
+// MARK: Help
+extension Text {
+    enum Help: String, Localizable, CaseIterable {
+        
+        case help = "Help"
+        case indexOfTip = "{{tipIndex}} of {{totalTipCount}} tips"
+        
+        enum Home: String, Localizable, CaseIterable {
+            case title = "See posts from users and topics you follow"
+            case body = "If your feed is empty, open the Discover tab, look for something interesting and follow users or topics to see their posts in your Home Feed."
+            case highlightedWord = "Discover"
+        }
+        
+        enum Discover: String, Localizable, CaseIterable {
+            case title = "See what's new and grow your network"
+            case body = "See posts from users and topics followed by your friends. Follow them to see their content in your Home Feed!"
+            case highlightedWord = "Home Feed"
+        }
+        
+        enum Notifications: String, Localizable, CaseIterable {
+            case title = "Keep up-to-date in all your conversations"
+            case body = "Check this screen for replies and reactions to your posts as well as conversations you participate in. We'll also notify you when someone mentions you."
+        }
+        
+        enum Hashtags: String, Localizable, CaseIterable {
+            case title = "What the community is talking about"
+            case body = "Browse through thousands of topics, engage in conversations and keep up with the hottest topics with people on Your Network."
+            case highlightedWord = "Your Network"
+        }
+        
+        enum YourNetwork: String, Localizable, CaseIterable {
+            case title = "Your friends, connections and pubs"
+            case body = "Your network is unique: it's made by the users you follow, those who *they* follow and the Pub servers you use to *gossip* messages with all of them."
+            case highlightedWord = "*gossip*"
+        }
     }
 }